Missiles and Laser fixed
The prefabs check marked are the ones that are used.

From previous commit
  Ok I just fixed the missile and laser prefabs to the right color and
used the aim120 prefab.  PLEASE.  WHATEVER YOU DO, DO NOT DELETE, MOVE,
OR CHANGE the prefab 'aim120', 'AIM120', or 'Missile' prefab in BOTH
the hiearchy, project, and inspector!!!!  Don't work on the missile AT
ALL.  I'll deal with the missile.
[2/15/14 11:34:55 AM] Jordan Hua: All those copies and prefabs are
branched with eachother so if you delete, open, or even move one
around, the whole thing gets messed up.  I don't care if it is messy.
 DO NOT MOVE OR TOUCH any of the missile stuff.  AND, DO NOT touch the
PlayerShoots folder folder for whatever reason!!!!