Saved the player as a prefab
Everytime you guys finish making and debugging an object, make sure you
save it as a prefab in case something happens to it in hierarchy.  DO
not delete the version in hierarchy because that is still the original
one.  To make any changes to it, simply change the one in the
hierarchy.  And then reattach it to the prefab folder to save the most
updated version.