Attaching spawn point to player solves 2 problems.
Ok I think I just fixed the problem.  I attached spawnpoint to the
player so that way the next wave spawns wherever the player is at.
This also has a dual benefit additonally: Because the levels are so
big, we don't want all the enemies to spawn at one place, but we also
don't want to spread out the enemies on the map because then it would
be too hard to find the enemies.  So attaching the spawnpoint object to
the player solved 2 problems.  The other alternative would be not only
fixing the next wave to spawn at the place player teleported to, but
the minimap needs arrows pointing to location of each enemy since the
enemies will be spread out.  Solution of now:  attaching the spawnpoint
object to the player solved 2 problems.